fix(review-loop): iteration 3 - resolve 0 error(s) + 1 warning(s)
